fix(support): move to separate blueprint
We used to register error handlers for both 400 and `ValidationError`
responses on the **global** Flask UI application. Depending on app
initialization order this could very likely conflict with other error
handlers.

Now the support view is on its separate blueprint which properly scopes
error handling to it.
