+++ invenio.cfg
@@ -952,7 +952,7 @@ RATELIMIT_PER_ENDPOINT = {
     "invenio_github_badge.latest_doi": "120 per minute",
     "invenio_github_badge.latest_doi_old": "120 per minute",
     # Support
-    "zenodo_rdm.support_form": "10 per second",
+    "zenodo_support.index": "10 per second",
     # Login and registration
     "invenio_oauthclient.login": "10 per minute",
     "invenio_oauthclient.signup": "10 per minute",
